refactor: createGraph から filter と abstraction を分離
